pyproject.toml
[tool.poetry]
name = "policy-search"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "*"
click = "*"
pyyaml = "*"
tqdm = "*"
pymupdf = "*"
install = "*"
sortedcontainers = "*"
setuptools = "*"
wheel = "*"
spacy = "*"
boto3 = "*"
mangum = "*"
uvicorn = "*"
fastapi = "*"
elasticsearch = "*"



[tool.poetry.dev-dependencies]
jupyterlab = "^3.1.14"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
